Fix memory leack

+++ observer_C/include/interface_observer.h
@@ -1,7 +1,7 @@
 #ifndef _OBSERVER_H
 #define _OBSERVER_H
 
-typedef void (*update_display) (int id, int temp, int time);
+typedef void (*update_display) (int id, float temp, float time);
 
 typedef struct observer
 {
